pyproject.toml

[project]
name = "wtplan"
version = "0.1.0"
description = "Issue-driven multi-repo worktree manager with MCP stdio server"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "TBD" }
authors = [{ name = "wtplan contributors" }]
dependencies = [
  "mcp>=1.26.0",
  "PyYAML>=6.0.1",
  "rich>=13.7.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "pre-commit>=3.6.0",
]

[project.scripts]
wtplan = "wtplan.cli:main"

[tool.ruff]
line-length = 128
src = ["src"]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "C4", "N", "PL", "RUF"]
ignore = ["PLR0913", "PLR0912", "RUF001", "RUF002", "PLR5501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]
